simplify names
