--- src/MQTTnet.Extensions.MultiCloud.BrokerIoTClient/WritablePropertyUTFJson.cs
@@ -1,19 +1,0 @@
-﻿using MQTTnet.Client;
-using MQTTnet.Extensions.MultiCloud.Binders;
-using MQTTnet.Extensions.MultiCloud.Serializers;
-
-namespace MQTTnet.Extensions.MultiCloud.BrokerIoTClient
-{
-    public class WritablePropertyUTFJson<T> : CloudToDeviceBinder<T, Ack<T>>, IWritableProperty<T>
-    {
-        public T? Value { get; set; }
-        public int? Version { get; set; }
-
-        public WritablePropertyUTFJson(IMqttClient c, string name)
-            : base(c, name, new UTF8JsonSerializer())
-        {
-            TopicTemplate = "device/{clientId}/props/{name}/set";
-            ResponseTopic = "device/{clientId}/props/{name}/ack";
-        }
-    }
-}
