+++ tests/MQTTnet.Extensions.MultiCloud.UnitTests/HubClient/HubTelemetryUTF8JsonFixture.cs
@@ -10,7 +10,7 @@ namespace MQTTnet.Extensions.MultiCloud.UnitTests.HubClient
         public async Task SendTelemetry()
         {
             var mqttClient = new MockMqttClient();
-            var telemetryBinder = new HubTelemetryUTF8Json<int>(mqttClient, "temp");
+            var telemetryBinder = new Telemetry<int>(mqttClient, "temp");
             await telemetryBinder.SendMessageAsync(2);
             Assert.Equal("devices/mock/messages/events/", mqttClient.topicRecceived);
             Assert.Equal("{\"temp\":2}", mqttClient.payloadReceived);
